rename model to stream
