+++ src/model/stream.rs
